backend/requirements.txt
anyio==3.6.2
attrs==22.2.0
bcrypt==4.0.1
beanie==1.17.0
build==0.10.0
CacheControl==0.12.11
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cleo==2.0.1
click==8.1.3
configparser==5.3.0
crashtest==0.4.1
cryptography==39.0.1
distlib==0.3.6
dnspython==2.3.0
dulwich==0.21.3
ecdsa==0.18.0
email-validator==1.3.1
et-xmlfile==1.1.0
fastapi==0.89.1
filelock==3.9.0
h11==0.14.0
html5lib==1.1
httptools==0.5.0
idna==3.4
importlib-metadata==6.0.0
installer==0.6.0
jaraco.classes==3.2.3
jeepney==0.8.0
jsonschema==4.17.3
keyring==23.13.1
lazy-model==0.0.5
libs==0.0.10
lockfile==0.12.2
more-itertools==9.1.0
motor==3.1.1
msgpack==1.0.4
multidict==6.0.4
npm==0.1.1
npx==0.1.1
numpy==1.24.2
openpyxl==3.1.1
optional-django==0.1.0
packaging==23.0
pandas==1.5.3
passlib==1.7.4
pexpect==4.8.0
pkginfo==1.9.6
platformdirs==2.6.2
poetry==1.4.0
poetry-core==1.5.1
poetry-plugin-export==1.3.0
psutil==5.9.4
ptyprocess==0.7.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.5
pydicom==2.3.1
pymongo==4.3.3
pyproject_hooks==1.0.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-decouple==3.7
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.7.1
PyYAML==6.0
rapidfuzz==2.13.7
requests==2.28.2
requests-toolbelt==0.10.1
rsa==4.9
SecretStorage==3.3.3
shellingham==1.5.0.post1
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
termcolor==2.2.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
trove-classifiers==2023.2.20
typing_extensions==4.5.0
urllib3==1.26.14
uuid==1.30
uvicorn==0.20.0
uvloop==0.17.0
virtualenv==20.20.0
watchfiles==0.18.1
webencodings==0.5.1
websockets==10.4
yarl==1.8.2
zipp==3.15.0
